refactor #step2 - 클래스 이름을 변경합니다
